backend/requirements.txt
flask
flask-cors
requests
transformers
torch
pillow
